+++ lib/Dialect/TorchConversion/Transforms/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_mlir_library(TorchMLIRTorchConversionPasses
   BackendTypeConversion.cpp
+  BackendTypeConversionPasses.cpp  
   Passes.cpp
   VerifyInvariantsBeforeBackendLowering.cpp
   VerifyLinalgOnTensorsBackendContract.cpp
